+++ miniconf_derive/src/lib.rs
@@ -2,7 +2,6 @@ use proc_macro::TokenStream;
 use quote::quote;
 use syn::{parse_macro_input, DeriveInput};
 
-mod attributes;
 mod field;
 
 use field::StructField;