+++ src/lib.rs
@@ -1,39 +1,31 @@
 #![cfg_attr(not(any(test, doctest)), no_std)]
-#![doc = include_str!("../README.md")]
+#![cfg_attr(feature = "json", doc = include_str!("../README.md"))]
 
 use core::fmt::Write;
 
-mod array;
-mod iter;
-mod option;
-
-pub use array::Array;
-pub use iter::MiniconfIter;
 pub use miniconf_derive::Miniconf;
-pub use option::Option;
+mod array;
+pub use array::*;
+mod iter;
+pub use iter::*;
+mod option;
+pub use option::*;
 
-#[cfg(feature = "mqtt-client")]
+#[cfg(feature = "json")]
+mod json;
+#[cfg(feature = "json")]
+pub use json::*;
+
+#[cfg(feature = "mqtt")]
+pub use minimq; // re-export
+#[cfg(feature = "mqtt")]
 mod mqtt_client;
+#[cfg(feature = "mqtt")]
+pub use mqtt_client::*;
 
-#[cfg(feature = "mqtt-client")]
-pub use mqtt_client::MqttClient;
-
-// Re-exports
-pub use heapless;
-pub use serde;
-pub use serde_json_core;
-
-#[cfg(feature = "mqtt-client")]
-pub use minimq;
-
-#[cfg(feature = "mqtt-client")]
-pub use minimq::embedded_time;
-
+pub use serde; // re-export
 #[doc(hidden)]
-pub use serde::{
-    de::{Deserialize, DeserializeOwned},
-    ser::Serialize,
-};
+pub use serde::{de::DeserializeOwned, Serialize};
 
 /// Errors that occur during iteration over topic paths.
 #[non_exhaustive]
@@ -247,50 +239,4 @@ pub trait SerDe<S>: Miniconf + Sized {
     /// # Returns
     /// The number of bytes used in the `data` buffer or an [Error].
     fn get(&self, path: &str, data: &mut [u8]) -> Result<usize, Error<Self::SerError>>;
-}
-
-/// Marker struct for the "JSON and `/`" [SerDe] specification.
-///
-/// Access items with `'/'` as path separator and JSON (from `serde-json-core`)
-/// as serialization/deserialization payload format.
-pub struct JsonCoreSlash;
-
-impl<T> SerDe<JsonCoreSlash> for T
-where
-    T: Miniconf,
-{
-    const SEPARATOR: char = '/';
-    type DeError = serde_json_core::de::Error;
-    type SerError = serde_json_core::ser::Error;
-
-    fn set(&mut self, path: &str, data: &[u8]) -> Result<usize, Error<Self::DeError>> {
-        let mut de = serde_json_core::de::Deserializer::new(data);
-        self.set_path(&mut path.split(Self::SEPARATOR).skip(1), &mut de)?;
-        de.end().map_err(Error::PostDeserialization)
-    }
-
-    fn get(&self, path: &str, data: &mut [u8]) -> Result<usize, Error<Self::SerError>> {
-        let mut ser = serde_json_core::ser::Serializer::new(data);
-        self.get_path(&mut path.split(Self::SEPARATOR).skip(1), &mut ser)?;
-        Ok(ser.end())
-    }
-}
-
-// These allow unifying serde error information to make writing examples
-// and tests easier. Doing this conversion is optional.
-// #[cfg(any(test, doctest))]
-impl From<Error<serde_json_core::ser::Error>> for Error<serde_json_core::de::Error> {
-    fn from(value: Error<serde_json_core::ser::Error>) -> Self {
-        match value {
-            Error::BadIndex => Self::BadIndex,
-            Error::PathAbsent => Self::PathAbsent,
-            Error::PathNotFound => Self::PathNotFound,
-            Error::PathTooLong => Self::PathTooLong,
-            Error::PathTooShort => Self::PathTooShort,
-            Error::PostDeserialization(_) => {
-                Error::PostDeserialization(serde_json_core::de::Error::CustomError)
-            }
-            Error::SerDe(_) => Self::SerDe(serde_json_core::de::Error::CustomError),
-        }
-    }
 }