+++ src/mqtt_client/mod.rs
@@ -1,14 +1,12 @@
-use serde_json_core::heapless::{String, Vec};
-
 use crate::{JsonCoreSlash, Miniconf, SerDe};
+use core::fmt::Write;
 use minimq::{
     embedded_nal::{IpAddr, TcpClientStack},
     embedded_time,
     types::{SubscriptionOptions, TopicFilter},
     Publication, QoS, Retain,
 };
-
-use core::fmt::Write;
+use serde_json_core::heapless::{String, Vec};
 
 // The maximum topic length of any settings path.
 const MAX_TOPIC_LENGTH: usize = 128;