+++ tests/iter.rs
@@ -1,4 +1,6 @@
-use miniconf::{heapless::String, Miniconf, SerDe};
+#![cfg(feature = "json")]
+
+use miniconf::{Miniconf, SerDe};
 
 #[derive(Miniconf, Default)]
 struct Inner {
@@ -21,7 +23,7 @@ fn insufficient_space() {
     assert_eq!(meta.count, 3);
 
     // Ensure that we can't iterate if we make a state vector that is too small.
-    assert!(Settings::iter_paths::<1, String<256>>().is_err());
+    assert!(Settings::iter_paths::<1, String>().is_err());
 }
 
 #[test]
@@ -32,7 +34,7 @@ fn test_iteration() {
         ("/c/inner".to_string(), false),
     ]);
 
-    for field in Settings::iter_paths::<32, String<256>>().unwrap() {
+    for field in Settings::iter_paths::<32, String>().unwrap() {
         assert!(iterated.contains_key(&field.as_str().to_string()));
         iterated.insert(field.as_str().to_string(), true);
     }
@@ -51,7 +53,7 @@ fn test_array_iteration() {
 
     let mut settings = Settings::default();
 
-    for field in Settings::iter_paths::<32, String<256>>().unwrap() {
+    for field in Settings::iter_paths::<32, String>().unwrap() {
         settings.set(&field, b"true").unwrap();
     }