+++ tests/structs.rs
@@ -1,4 +1,6 @@
-use miniconf::{heapless::String, Miniconf, SerDe};
+#![cfg(feature = "json")]
+
+use miniconf::{Miniconf, SerDe};
 use serde::{Deserialize, Serialize};
 
 #[test]
@@ -97,7 +99,7 @@ fn struct_with_string() {
 fn empty_struct() {
     #[derive(Miniconf, Default)]
     struct Settings {}
-    assert!(Settings::iter_paths::<1, String<0>>()
+    assert!(Settings::iter_paths::<1, String>()
         .unwrap()
         .next()
         .is_none());